requirements.txt
# ARGUS Dependencies

# Core
python-dotenv==1.0.0
requests==2.31.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Real-time / WebSockets
websockets==12.0

# CLI / UI (Future)
rich==13.7.0
textual==0.47.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
